+++ torch_brain/models/calcium_poyo_plus.py
@@ -12,15 +12,17 @@ from torch_brain.data import chain, pad8, track_mask8
 from torch_brain.nn import (
     Embedding,
     InfiniteVocabEmbedding,
-    MultitaskReadout,
     RotaryCrossAttention,
     RotarySelfAttention,
     RotaryTimeEmbedding,
-    prepare_for_multitask_readout,
 )
 from torch_brain.registry import ModalitySpec, MODALITY_REGISTRY
-from torch_brain.models.poyo_plus import FeedForward
+from torch_brain.models.poyo_plus.poyo_plus import FeedForward
 
+from torch_brain.models.poyo_plus.multitask_readout import (
+    MultitaskReadout,
+    prepare_for_multitask_readout,
+)
 from torch_brain.utils import (
     create_linspace_latent_tokens,
     create_start_end_unit_tokens,
