+++ torch_brain/models/poyo_plus/__init__.py
@@ -1,0 +1,3 @@
+from .poyo_plus import POYOPlus, FeedForward
+
+__all__ = ["POYOPlus", "FeedForward"]
